Add files via upload
This code is a very basic one which only considers about 1 team in 1 match. However, it can be a start and foundation of the code which considers multiple teams in multiple games to analyze what we want: the strategies and features of the teams. We can use these to adopt targeted tactics against opponents.
Still have something to go further, like "Expected Threat".